+++ changelog.md
@@ -5,6 +5,9 @@
 - Add elipsis to event time listing to prevent overflow of text
 - Favorites are now grouped by the day they happen and are split based on occurrence
 - Fix for event showing the first occurrence not the favorited one
+- Display badge for "now" when an event is happening now
+- Favorite events that have passed do not show in the list
+- Fix to correctly choose todays day if its available in events
 
 ### 1.1 - August 2, 2023
 - Submitted to the App Store